+++ src/lib/components/ExpertsView.tsx
@@ -1,20 +1,25 @@
 import React, { useState } from 'react';
-
-import { Container, Typography } from '@material-ui/core';
+import {
+  CircularProgress,
+  Container,
+  Grid,
+  Typography,
+} from '@material-ui/core';
 import { ExpertBlock } from './ExpertBlock';
 import { useStyles } from '../../modules/main/styles/MainExpertsView.styles';
 import { ExpertPopover } from './ExpertPopover';
 import ExpertDataCard from './ExpertDataCard';
-import { IExpert } from '../types';
+import { IExpert, LoadingStatusEnum } from '../types';
 
 const cardsClasses = Array.from(Array(11).keys()).map((el) => `item_${el}`);
 
 export interface IExpertsViewProps {
   cards: IExpert[];
+  loading?: LoadingStatusEnum;
 }
 
 export const ExpertsView: React.FC<IExpertsViewProps> = (props) => {
-  const { cards } = props;
+  const { cards, loading } = props;
   const [anchorEl, setAnchorEl] = useState<HTMLDivElement | null>(null);
   const [popoverCard, setPopoverCard] = useState<IExpert | null>(null);
 
@@ -43,9 +48,19 @@ export const ExpertsView: React.FC<IExpertsViewProps> = (props) => {
     </div>
   ));
 
+  const renderError = (errorMsg = 'Не вдалося завантажити експертів') =>
+    loading === LoadingStatusEnum.failed ? <span>{errorMsg}</span> : null;
+
+  const renderLoading = () =>
+    loading === LoadingStatusEnum.pending ? <CircularProgress /> : null;
+
   return (
     <Container>
       <Typography variant="h4">Експерти</Typography>
+      <Grid container direction="column" alignItems="center">
+        {renderLoading()}
+        {renderError()}
+      </Grid>
       <div className={classes.container}>
         {allExperts}
         <ExpertPopover
